Relocate file_name_for_content_block to util.taxii
